split routes into their own files
also version API and update tests.

This change should allow making version changes to the API easier and
also makes the code a bit more modular and so easier to read and find
things.
